feature: Adding profile to execute the application with profile "dev" and "test"
