+++ src/main/java/com/example/WeibisWeb/dao/FileUpLoadedRepository.java
@@ -1,6 +1,7 @@
 package com.example.WeibisWeb.dao;
 
 import com.example.WeibisWeb.resources.UploadedFile;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +12,7 @@ import java.util.UUID;
  * Repository layer of File Uploaded
  */
 @Repository
+@Profile(value = {"dev", "test"})
 public interface FileUpLoadedRepository extends JpaRepository<UploadedFile, UUID> {
 
     /**
