+++ src/main/java/com/example/WeibisWeb/service/JobDescriptionServiceImpl.java
@@ -9,6 +9,7 @@ import com.example.WeibisWeb.exception.JobDescriptionNotFoundException;
 import com.example.WeibisWeb.resources.JobDescription;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -25,6 +26,7 @@ import java.util.stream.Collectors;
 @Service
 @Transactional
 @RequiredArgsConstructor
+@Profile(value = {"dev", "test"})
 public class JobDescriptionServiceImpl implements JobDescriptionService {
 
     private final JobDescriptionRepository jobDescriptionRepository;
