+++ src/main/resources/application-dev.properties
@@ -1,0 +1,9 @@
+# Memory Database for development Environment
+spring.h2.console.enabled=true
+spring.he.console.path=/h2-console
+spring.datasource.url=jdbc:h2:mem:devdb;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.jpa.hibernate.hibernate.dialect=org.hibernate.dialect.H2Dialect
+spring.jackson.serialization.fail-on-empty-beans=false
