+++ src/main/resources/application.properties
@@ -1,16 +1,6 @@
 # The server port
 server.port=8081
 
-# Memory Database
-spring.h2.console.enabled=true
-spring.he.console.path=/h2-console
-spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_ON_EXIT=FALSE
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jackson.serialization.fail-on-empty-beans=false
-
 # Setting the size of a file
 spring.servlet.multipart.enabled=true
 spring.servlet.multipart.file-size-threshold=2KB
@@ -25,3 +15,7 @@ spring.cache.redis.time-to-live=600000
 spring.cache.cache-names=Users
 spring.cache.redis.key-prefix=jdj
 spring.data.redis.repositories.type = false
+
+
+# Set the profile to switch environments
+spring.profiles.active=dev
